WIP: Generalise type of CRTInfo
When returning the i'th power for some m'th root of unity, we must be able to parameterise over the type 'i'. For the Accelerate backend this is going to be some embedded expression type (Exp Int) rather than a real Int.

This isn't completely done yet; in particular, the Tensor class just forces this to 'Int' so that everything continues to work as previously, rather than that leaving that free.