Split test suite into a separate package
The motivation for this split being so that we can also use this to test the Accelerate backend. Unfortunately, we lose the 'cabal test' command.

Note that there are some symlinks back to the common 'utils' files. Once the benchmark suite is updated and split off as well, decide whether to duplicate these between each, or continue to have both point to a common location.